More warnings removal

+++ source/unix/unix_console.cpp
@@ -46,7 +46,7 @@ char *Sys_ConsoleInput( void ) {
 
 static void Sys_AnsiColorPrint( const char *msg ) {
 	static char buffer[2096];
-	int length = 0;
+	size_t length = 0;
 	static const int q3ToAnsi[ MAX_S_COLORS ] =
 	{
 		30, // COLOR_BLACK
